♻️ refactor: 콜렉션 페이지 컴포넌트 이동 - /components/collect -> /features/collect/components
