--- src/components/collect/Collection/index.ts
@@ -1,1 +1,0 @@
-export * from "./Collection";
