--- src/components/collect/SearchedCollection/index.ts
@@ -1,1 +1,0 @@
-export * from "./SearchedCollection";
