+++ src/features/collect/components/SearchedCollection.tsx
