+++ src/pages/collect/index.tsx
@@ -5,9 +5,9 @@ import { BackButtonNavigation } from "@/common/components/Navigation";
 import { MemeListSkeleton } from "@/common/components/Skeleton";
 import { SSRSuspense } from "@/common/components/Suspense";
 import { useDebounce, useInput } from "@/common/hooks";
-import { Collection, SearchedCollection } from "@/components/collect";
 import { withAuth } from "@/components/hocs";
 import { SearchInput } from "@/components/search";
+import { Collection, SearchedCollection } from "@/features/collect/components";
 
 const CollectPage = () => {
   const inputProps = useInput();
